Implement PR review fixes for Solana client
- Extract timeout constant (DEFAULT_TRANSACTION_TIMEOUT_SECONDS)
- Rename SolanaSigningClientImpl → SolanaClient
- Move transfer_sol_amount and airdrop_sol_amount to SolanaSigningClient trait
- Add mnemonic-based constructors using BIP39 derivation (m/44'/501'/0'/0')
- Remove get_keypair redundancy from SolanaLocalnetClient
- Make SolanaTestClient configurable with custom RPC URLs
- Update all tests and trait implementations

Addresses all technical feedback from PR review:
- Fixes hardcoded timeout values
- Improves API organization and consistency
- Adds mnemonic support following cosmos client patterns
- Removes network-specific code distinctions
- Maintains backward compatibility
